refactor: change the file organization
