+++ src/components/Footer/Footer.css
@@ -1,0 +1,38 @@
+.footer {
+  bottom: 0;
+  padding: 68px 0 60px;
+  display: flex;
+  justify-content: flex-start;
+  margin: 0 auto;
+}
+
+@media screen and (max-width: 919px) {
+  .footer {
+    max-width: 580px;
+  }
+}
+
+@media screen and (max-width: 619px) {
+  .footer {
+    max-width: 280px;
+  }
+}
+
+@media screen and (max-width: 375px) {
+  .footer {
+    padding: 48px 0 36px;
+  }
+}
+
+.footer__copyright {
+  font-size: 18px;
+  line-height: 22px;
+  color: #545454;
+  margin: 0;
+}
+
+@media screen and (max-width: 619px) {
+  .footer__copyright {
+    font-size: 14px;
+  }
+}
