+++ src/components/Footer/Footer.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import './Footer.css';
+
+function Footer() {
+  return (
+    <footer className="footer__position">
+      <div className="footer">
+        <p className="footer__copyright">&copy; 2020 Mesto Russia</p>
+      </div>
+    </footer>
+  );
+}
+
+export default Footer;
